changes in clinic admin

+++ clinicadmin/src/main/java/com/clinicadmin/service/impl/TherophyProgramServiceImpl.java
@@ -186,7 +186,7 @@ public class TherophyProgramServiceImpl implements TherophyProgramService {
                     List<TheraphyNamesDTO> therapyList =
                             new LinkedList<>();
 
-                    double totalProgramAmount = 0;
+                    int totalProgramAmount = 0;
 
                     if (e.getTherophyIds() != null &&
                             !e.getTherophyIds().isEmpty()) {
